+++ src/components/block/index.ts
@@ -201,7 +201,7 @@ export default class Block extends EventsDispatcher<BlockEvents> {
   /**
    * Is fired when DOM mutation has been happened
    */
-  private didMutated = _.debounce((mutations: MutationRecord[]): void => {
+  private didMutated = _.debounce((mutations: MutationRecord[] = []): void => {
     const shouldFireUpdate = !mutations.some(({ addedNodes = [], removedNodes }) => {
       return [...Array.from(addedNodes), ...Array.from(removedNodes)]
         .some(node => $.isElement(node) && (node as HTMLElement).dataset.mutationFree === 'true');
@@ -700,6 +700,14 @@ export default class Block extends EventsDispatcher<BlockEvents> {
   public willUnselect(): void {
     this.mutationObserver.disconnect();
     this.removeInputEvents();
+  }
+
+  /**
+   * Allows to say Editor that Block was changed. Used to manually trigger Editor's 'onChange' callback
+   * Can be useful for block changes invisible for editor core.
+   */
+  public dispatchChange(): void{
+    this.didMutated();
   }
 
   /**